test: establish independent core contracts and state-sequence gate
