+++ packages/core/package.json
@@ -41,7 +41,17 @@
     "unist-util-visit": "^5.1.0"
   },
   "devDependencies": {
-    "@types/node": "^25.9.5"
+    "@types/node": "^25.9.5",
+    "unified": "^11.0.5",
+    "remark-parse": "^11.0.0",
+    "remark-gfm": "^4.0.1",
+    "remark-rehype": "^11.1.2",
+    "remark-squeeze-paragraphs": "^6.0.0",
+    "remark-remove-comments": "^1.1.1",
+    "@ai-markdown/rehype-raw": "7.0.3",
+    "rehype-sanitize": "^6.0.0",
+    "vfile": "^6.0.3",
+    "fast-check": "^4.9.0"
   },
   "publishConfig": {
     "access": "public"
